+++ satella/os/signals.py
@@ -1,29 +1,29 @@
 import signal
 import time
+import logging
 
 import typing as tp
 
+logger = logging.getLogger(__name__)
 end = False
 
 
 def __sighandler(a, b):
+    logger.warning('Handling a signal')
     global end
     end = True
 
 
-def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None,
-                   end_on_keyboard_interrupt: bool = True) -> None:
+def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None) -> None:
     """
     Will hang until this process receives SIGTERM or SIGINT.
     If you pass extra signal IDs (signal.SIG*) with extra_signals,
     then also on those signals this call will release.
 
     :param extra_signals: a list of extra signals to listen to
-    :param end_on_keyboard_interrupt: whether to consider receiving a KeyboardInterrupt as
-        a signal to finish
     """
     global end
-    extra_signals = extra_signals or []
+    extra_signals = extra_signals or ()
 
     signal.signal(signal.SIGTERM, __sighandler)
     signal.signal(signal.SIGINT, __sighandler)
@@ -31,10 +31,6 @@ def hang_until_sig(extra_signals: tp.Optional[tp.List[int]] = None,
         signal.signal(s, __sighandler)
 
     while not end:
-        try:
-            time.sleep(0.5)
-        except KeyboardInterrupt:
-            if end_on_keyboard_interrupt:
-                end = True
+        time.sleep(0.5)
 
     end = False  # reset for next use