+++ tests/test_os/test_hang_until_sig.py
@@ -3,8 +3,8 @@ import signal
 import sys
 import threading
 import unittest
+import time
 
-from satella import time
 from satella.os import hang_until_sig
 
 
@@ -12,10 +12,10 @@ class TestHangUntilSig(unittest.TestCase):
 
     @unittest.skipIf('win' in sys.platform, 'Cannot test on Windows')
     def test_hang_until_sig(self):
+
         def send_sig():
             time.sleep(1)
-            os.kill(os.getpid(), signal.SIGINT)
+            os.kill(0, signal.SIGTERM)
 
         threading.Thread(target=send_sig).start()
         hang_until_sig()
-        self.assertEqual(True, False)